Split this test into its original parts
